work on making it run

+++ pasta.py
@@ -93,9 +93,9 @@ class PastaEncoder(Model):
         data_sorted = Dataset(sorted(data_unsorted.instances, key=lambda instance: len(instance.fields['text'].tokens), reverse=True))
         lengths = [len(instance.fields['text'].tokens) for instance in data_sorted.instances]
         array_dict = data_sorted.as_array_dict()
-        embedded = self.text_field_embedder({
-            'tokens': Variable(array_dict['text']['tokens'], requires_grad=False),
-            'token_characters': Variable(array_dict['text']['token_characters'], requires_grad=False)
+        embedded = self.text_field_embedder({ # TODO fix trying to encode zero-length character sequences
+            'tokens': Variable(torch.LongTensor(array_dict['text']['tokens']), requires_grad=False),
+            'token_characters': Variable(torch.LongTensor(array_dict['text']['token_characters']), requires_grad=False)
         })
         packed = pack_padded_sequence(embedded, lengths, batch_first=True)
         word_enc_out, word_enc_hidden = self.word_enc_lstm(packed)
